moved files outside frontend folder
