+++ .env
@@ -1,1 +1,3 @@
-REACT_APP_JWT_SECRET='secret123'
+REACT_APP_JWT_SECRET='secret123',
+FAST_REFRESH=false,
+CHOKIDAR_USEPOLLING=true
