+++ src/Pages/Cart/Cart.jsx
@@ -5,7 +5,7 @@ const Cart=()=>{
     return(
         <div>
          <Navbar/>
-         <h1>Cart page work in progress...</h1>
+         <h1>Cart page work in progress</h1>
         </div>
     )
