+++ src/Utils/filterByPrice.js
@@ -1,0 +1,6 @@
+export const filterByPrice = (products, filter) => {
+  const { priceRangeSlider } = filter;
+  return products.filter(
+    (product) => Number(product.price) <= priceRangeSlider
+  );
+};
