+++ src/components/ProductListingCard/ProductListingCard.css
@@ -1,0 +1,75 @@
+.vertical-card {
+  display: flex;
+  flex-direction: column;
+  max-width: 14rem;
+  border: 2px var(--border);
+  box-shadow: var(--shadow);
+  border-radius: 0.3rem;
+  margin: 1rem;
+  padding: 1rem;
+}
+.price {
+  margin-left: -0.3rem;
+}
+.product-image {
+  border-radius: 0.4rem;
+}
+.discount {
+  color: #10b981;
+}
+.btn-cta-vertical {
+  width: 100%;
+  /* font-size: 1.5rem; */
+  display: inline-block;
+  background-color: var(--cta);
+  color: var(--white);
+  border: none;
+  border-radius: 0.3rem;
+  margin-top: 0.5rem;
+  padding: 0.4rem 0;
+  cursor: pointer;
+}
+.wishlist-icon {
+  border: 2px solid var(--border);
+  border-radius: 50%;
+  padding: 0.5rem;
+  background-color: var(--white);
+  transform: translateX(8rem) translateY(-18px) scale(1) rotate(0deg)
+    translateZ(0px);
+}
+.wishlist-icon:hover {
+  border: 2px solid var(--border);
+  border-radius: 50%;
+  padding: 0.5rem;
+  background-color: var(--white);
+  transform: translateX(8rem) translateY(-18px) scale(1) rotate(20deg)
+    translateZ(0px);
+  cursor: pointer;
+}
+.title {
+  margin: 0rem;
+  display: flex;
+  justify-content: space-between;
+}
+.details {
+  margin-top: -2rem;
+}
+@media (max-width: 550px) {
+  .wishlist-icon {
+    border: 2px solid var(--border);
+    border-radius: 50%;
+    padding: 0.5rem;
+    background-color: var(--white);
+    transform: translateX(10rem) translateY(-18px) scale(1.2) rotate(0deg)
+      translateZ(0px);
+  }
+  .wishlist-icon:hover {
+    border: 2px solid var(--border);
+    border-radius: 50%;
+    padding: 0.5rem;
+    background-color: var(--white);
+    transform: translateX(10rem) translateY(-18px) scale(1.2) rotate(20deg)
+      translateZ(0px);
+    cursor: pointer;
+  }
+}
