--- src/frontend/Utils/filterByPrice.js
@@ -1,6 +1,0 @@
-export const filterByPrice = (products, filter) => {
-  const { priceRangeSlider } = filter;
-  return products.filter(
-    (product) => Number(product.price) <= priceRangeSlider
-  );
-};
