--- src/frontend/Utils/index.js
@@ -1,5 +1,0 @@
-import { filterByCategory } from "./filterByCategory.js";
-import { filterByRating } from "./filterByRating.js";
-import { sortByPrice } from "./sortByPrice";
-import { filterByPrice } from "./filterByPrice";
-export { filterByCategory, filterByRating, sortByPrice, filterByPrice };
