+++ src/index.js
@@ -4,9 +4,9 @@ import "./index.css";
 import { App } from "./App.js";
 import { makeServer } from "./server";
 import { BrowserRouter as Router } from "react-router-dom";
-import { DataProvider } from "../src/frontend/contexts/DataContext.jsx";
-import { ProductsProvider } from "../src/frontend/contexts/ProductContext.jsx";
-import { AuthProvider } from "../src/frontend/contexts/AuthContext";
+import { DataProvider } from "../src/contexts/DataContext.jsx";
+import { ProductsProvider } from "../src/contexts/ProductContext.jsx";
+import { AuthProvider } from "../src/contexts/AuthContext";
 // Call make Server
 makeServer();
